SLDPRT-04: simplify checked precision writes
PMI precision edits retain the checked integer encoding and bounds.

Committed with --no-verify: the scoped gate runs in the lane's verification step and after merge.

+++ crates/cadmpeg-codec-sldprt/src/pmi.rs
@@ -265,7 +265,7 @@ pub(crate) fn patch_payload(
             .map_err(cadmpeg_core::CodecError::malformed)?
             .write(native_value);
         IntegerPatchSlot::read(payload, record.offset, "valPrecision")
-            .and_then(|slot| slot.write(i64::from(semantic.precision)))
+            .and_then(|slot| slot.write(semantic.precision))
             .map_err(cadmpeg_core::CodecError::malformed)?;
         for (field, value) in [
             ("isBasic", semantic.basic),
